Refactored alchemy functions from menu
