+++ client/src/components/menu.jsx
@@ -1,13 +1,8 @@
 import { Link } from "react-router-dom";
 import { updateUser } from "../services/userServices";
-import { getItems, addItem } from "../services/itemServices";
-import {
-  getPotionTimer,
-  addPotionTimer,
-} from "../services/potionTimerServices";
+import { getPotionTimer } from "../services/potionTimerServices";
 import { useState } from "react";
 import { potionsList } from "../constants/items";
-import { ingredientsList } from "../constants/items";
 import recipeList from "../constants/recipes";
 
 function Menu({
@@ -41,13 +36,14 @@ function Menu({
   alchemyStatus,
   setAlchemyStatus,
   potions,
-  setPotions,
   ingredients,
-  setIngredients,
   summonHPBonus,
   setSummonHPBonus,
   summonMPBonus,
   setSummonMPBonus,
+  loadDataAlchemy,
+  createPotion,
+  consumePotion,
 }) {
   // numbered index state (summons, and recipes pagination)
   const [index1, setIndex1] = useState(0);
@@ -67,7 +63,6 @@ function Menu({
   const [potionsStatus, setPotionsStatus] = useState(false);
   const [ingredientsStatus, setIngredientsStatus] = useState(false);
   const [recipesStatus, setRecipesStatus] = useState(false);
-  const [potionCooldown, setPotionCooldown] = useState(false);
 
   // paginates creatures for summons menu
   const paginateCreatures = async (index1, direction) => {
@@ -294,221 +289,6 @@ function Menu({
       setBattleStatus(true);
       setBattleUndecided(true);
       await loadAsyncDataPlayer();
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  // loads alchemy data
-  const loadDataAlchemy = async () => {
-    try {
-      setPotions([]);
-      setIngredients([]);
-      const { data } = await getItems();
-      const playerPotionsData = data.filter(
-        (item) => item.type === "Potion" && item.userId === player.userfrontId
-      );
-      const playerPotions = potionsList.filter((potion) =>
-        playerPotionsData.some((item) => item.itemId === potion.id)
-      );
-      for (let i = 0; i < playerPotions.length; i++) {
-        playerPotions[i].itemQuantity = playerPotionsData.find(
-          (item) => item.itemId === playerPotions[i].id
-        ).itemQuantity;
-      }
-      setPotions(playerPotions);
-      const playerIngredientsData = data.filter(
-        (item) =>
-          item.type === "Ingredient" && item.userId === player.userfrontId
-      );
-      const playerIngredients = ingredientsList.filter((ingredient) =>
-        playerIngredientsData.some((item) => item.itemId === ingredient.id)
-      );
-      for (let i = 0; i < playerIngredients.length; i++) {
-        playerIngredients[i].itemQuantity = playerIngredientsData.find(
-          (item) => item.itemId === playerIngredients[i].id
-        ).itemQuantity;
-      }
-      setIngredients(playerIngredients);
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  // creates a new potion
-  const createPotion = async (potionId) => {
-    try {
-      if (!potionCooldown) {
-        setPotionCooldown(true);
-        const { data } = await getItems();
-        const playerPotionData = data.filter(
-          (item) => item.type === "Potion" && item.itemId === potionId
-        );
-        const potion = potionsList.find((item) => item.id === potionId);
-        const newPotionData = {
-          itemId: potion.id,
-          type: "Potion",
-          itemQuantity: playerPotionData[0]
-            ? playerPotionData[0].itemQuantity + 1
-            : 1,
-          userId: Userfront.user.userId,
-        };
-        const playerIngredientData = data.filter(
-          (item) => item.type === "Ingredient"
-        );
-        const currentRecipe = recipeList.filter(
-          (item) => item.potionProductId === potion.id
-        )[0];
-        // check if player has enough ingredients for recipe
-        const ingredient1Check = playerIngredientData.find(
-          (item) => item.itemId === currentRecipe.ingredient1
-        );
-        const ingredient2Check = playerIngredientData.find(
-          (item) => item.itemId === currentRecipe.ingredient2
-        );
-        if (
-          ingredient1Check &&
-          ingredient1Check.itemQuantity > 0 &&
-          ingredient2Check &&
-          ingredient2Check.itemQuantity > 0
-        ) {
-          // confirm potion creation
-          if (window.confirm(`Are you sure you want to create this potion?`)) {
-            const currentIngredient1 = playerIngredientData.find(
-              (ingredient) => ingredient.itemId === currentRecipe.ingredient1
-            );
-            const currentIngredient2 = playerIngredientData.find(
-              (ingredient) => ingredient.itemId === currentRecipe.ingredient2
-            );
-            // delete the ingredients from the player's inventory
-            currentIngredient1.itemQuantity -= 1;
-            currentIngredient2.itemQuantity -= 1;
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addItem(currentIngredient1);
-            await loadDataAlchemy();
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addItem(currentIngredient2);
-            await loadDataAlchemy();
-            // add the potion to the player's inventory
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addItem(newPotionData);
-            await loadDataAlchemy();
-            setTimeout(() => {
-              setPotionCooldown(false);
-            }, 1000);
-          } else {
-            setTimeout(() => {
-              setPotionCooldown(false);
-            }, 1000);
-          }
-        } else {
-          alert("You don't have enough ingredients for this potion.");
-          setTimeout(() => {
-            setPotionCooldown(false);
-          }, 1000);
-        }
-      }
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  // uses clicked potion
-  const consumePotion = async (potionId) => {
-    try {
-      if (!potionCooldown) {
-        await getPotionTimer();
-        const timerData = await getPotionTimer();
-        setPotionCooldown(true);
-        const { data } = await getItems();
-        const playerPotionData = data.filter(
-          (item) => item.type === "Potion" && item.itemId === potionId
-        );
-        const potion = potionsList.find((item) => item.id === potionId);
-        const currentPotionData = playerPotionData.filter(
-          (item) => item.itemId === potionId
-        );
-        // prevent over use
-        if (timerData.data.length !== 0) {
-          alert(
-            "You already have an active potion. Please wait for it to expire."
-          );
-          setTimeout(() => {
-            setPotionCooldown(false);
-          }, 1000);
-          return;
-        }
-        // check if player has enough potions
-        if (currentPotionData[0].itemQuantity > 0) {
-          // confirm potion use
-          if (
-            window.confirm(`Are you sure you want to use this potion?`) === true
-          ) {
-            // use potion
-            currentPotionData[0].itemQuantity -= 1;
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addItem(currentPotionData[0]);
-            await loadDataAlchemy();
-            await Userfront.user.update({
-              data: {
-                userkey: Userfront.user.data.userkey,
-              },
-            });
-            await addPotionTimer({
-              userId: Userfront.user.userId,
-              potionId: potion.id,
-              potionDuration: potion.duration,
-            });
-            await loadDataAlchemy();
-            // checks potion timer
-            const potionTimer = await getPotionTimer();
-            if (potionTimer.data.length > 0) {
-              const playerPotion = potionsList.find(
-                (potion) => potion.id === potionTimer.data[0].potionId
-              );
-              const playerMPBonus = playerPotion.mpMod;
-              const playerHPBonus = playerPotion.hpMod;
-              setSummonMPBonus(playerMPBonus);
-              setSummonHPBonus(playerHPBonus);
-            }
-
-            setPlayerCreatureMP(
-              playerCreature.mp + chosenRelic.mpMod + summonMPBonus
-            );
-            setPlayerCreatureHP(
-              playerCreature.hp + chosenRelic.hpMod + summonHPBonus
-            );
-            setTimeout(() => {
-              setPotionCooldown(false);
-            }, 1000);
-          } else {
-            setTimeout(() => {
-              setPotionCooldown(false);
-            }, 1000);
-          }
-        } else {
-          alert("You don't have enough potions.");
-          setTimeout(() => {
-            setPotionCooldown(false);
-          }, 1000);
-        }
-      }
     } catch (error) {
       console.log(error);
     }
